Introduce initialPoll()
When the axis is created, configuration values that need to be fetched
from the controller.
These are put into the paramater library and/or the record fields.
The most promiment example are the "read only soft limits".

Since different controllers share the same principle, add a generic
initialPoll() here in axisCore.