axisRecord: maybeRetry() needs the commanded position
There was a possible race condition, when RTRY == 1 and the
DVAL field was changed while the motor was moving.

At the end of the move, the DRBV field was compared to DVAL,
and this didn't work as expected.
Because the DVAL had changed under the time, the record did set
the "MISS" field.

Steps to reproduce:
Tweek the motor a couple of times fast after each other.

Introduse a private variable last.commandedDval, and compare it
with DRBV at the end of the move.