init_record() Controller offline error handling, part 1
When the record is initialized, the current postion of the axis is read
from the controller. If the position is 0 and if "auto save and restore"
for this axis is configured, the position will be set into the controller.

If the communication is not established (yet) this all fails and the
record should be put into an error state.

Do the first part of error handling:
- asynAxisController::readGenericPointer() checks if values like the position
  had been retrieved for the axis and returns asynError if not.
- If asyError is returned devAxisAsyn/init_record() sets the pact field to 1
  and the record stays in an UDF alarm.

Note:
  In the future, it may be possible to recover later, this needs
  refactoring of code: The parts of the initialization need to
  be broken out into seperate functions.