Axis not homed: May be "W: or "E:"
Axis not homed can either be ignored, it may be a warning or an error.
(Warning means that motorMessageText_ is set to "W: Axis not homed")
(Error means that motorMessageText_ is set to "E: Axis not homed")

The driver can configure this, by calling
setIntegerParam(pC_->motorNotHomedProblem_, motorNotHomedProblem);

motorNotHomedProblem should be one of these:
MOTORNOTHOMEDPROBLEM_IGNORE
MOTORNOTHOMEDPROBLEM_WARNING
MOTORNOTHOMEDPROBLEM_ERROR

Independent of the motorMessageText the driver must report a problem to
put the record into the alarm state.