axisRecord: Change mip with macros
Introduce mipSetBit() and mipClrBit() to set or clear single bits in the mip
field.
Add mipSetMip() to set the mip field to a new value.

All these functions are hidden behid macros:
MIP_SET_BIT(), MIP_CLR_BIT() and MIP_SET_VAL(), which will may be compiled with
or with a debug print, depending on #ifdef MIPDEBUG

This is to allow to debug a bug: When jogging and the motor is stopped by
switching CNEN to 0, the JOGF/JOGR field does not go to 0.
This will be fixed in the next commit.