axisRecord: Reset JOGF/JOGR when jogging stops (part 2)
When the motor is jogging and stopped by setting CNEN to 0, JOGF/JOGR may
stay at 1.
(This assumes that CNEN controls the amplifier, as it does in the simulator
and test 141).

Correct an over-critical check: pmr->mip may be either
exactly MIP_JOGF or exactly MIP_JOGR.
After this change the MIP_JOG_REQ bit may be 1 (or 0).
The
"Motor stopped while jogging and we didn't stop it"
code is executed, the movement is done, no backlash.