init_record() Controller offline error handling, part 2
This was changed in bbdce8ce077e44f63547779ee3e53f03844f:
>  When the record is initialized, the current postion of the axis is read
>  from the controller. If the position is 0 and if "auto save and restore"
>  for this axis is configured, the position will be set into the controller.

>  If the communication is not established (yet) this all fails and the
>  record should be put into an error state.

Improve the situation, if "auto save and restore" is not needed:
- Introduce  load_pos_needed() in devAxisAsyn.c, break code out from
  init_controller_load_pos_if_needed()
- Introduce initial_poll(), break out code from init_record()

Look at the return value of process_reason = (*pdset->update_values)
in axisrecord.cc and change the processing:

If the initial poll fails and "load position" is not needed,
put the record into UDF state.
After the first poll, move it out of the UDF state.

If the initial poll fails and "load position" -is- needed,
put the record into UDF state and set PACT to 1, making
the record unusable. The IOC must be restarted.

This improves the situation when auto save & restore is not used.
On the log run, all this initializion code should become more
asyn-ish and be part of the poller. Stay tuned.